agents/2-agentic-workflows/requirements.txt
jupyterlab
ipykernel
ipywidgets
openai-agents
openai>=1.12.0
python-dotenv>=1.0.0
pydantic>=2.0.0
google-api-python-client
google_auth_oauthlib